requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2==2.9.6
pycodestyle==2.10.0
pypng==0.20220715.0
pytz==2023.3
qrcode==7.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
